refactor(validation): move govalidator rules and struct validation into models
